// src/redux/features/game/types.ts
export interface Game {
    id: number;
    title: string;
    thumbnail: string;
    short_description: string;
    game_url: string;
    genre: string;
    platform: string;
    publisher: string;
    developer: string;
    release_date: string;
    freetogame_profile_url: string;
}

export interface GameSliceState {
    games: Game[];
    filteredGames: Game[];
    loading: boolean;
    error: string | null;
    selectedFilter: string;
}